(feat) Implement delete and update functionality for procedures

+++ client/app/procedures/procedures.js
@@ -7,21 +7,55 @@ angular.module('producer.procedures', ['alertMessage'])
   $scope.showProcedure = false;
 
   $scope.procedure = {};
-  $scope.newProcedure = {};
+  $scope.confirm = undefined;
 
   $scope.getClass = function (procedure) {
-    if(procedure === undefined || $scope.selected === undefined) {
+    if(procedure === {} || $scope.selected === undefined) {
       return '';
     }
     return Hilighter.hilight(procedure.title, $scope.selected.title);
   };
 
+  $scope.getProcedures = function(){
+    Procedures.getProcedures().then(function(response){
+      $scope.procedures = response.data.map(function(procedure) {
+        procedure.originalTitle = procedure.title;
+        return procedure;
+      });
+
+      // if existing procedure is currently selected
+      if ($scope.procedure.id) {
+        $scope.procedures.forEach(function(procedure) {
+          if (procedure.id === $scope.procedure.id) {
+            $scope.procedure = procedure;
+          }
+        });
+      }
+    });
+  };
+
+  $scope.setMaster = function(section) {
+    $scope.selected = $scope.procedure = section;
+    $scope.showProcedure = true;
+    $scope.confirm = undefined;
+    Messages.clearMessage();
+  };
+
+  $scope.createNewProcedure = function(){
+    $scope.procedure = {};
+    $scope.showProcedure = false;
+    $scope.selected = undefined;
+  };
+
   $scope.submitProcedure = function(){
-    Procedures.submitProcedure($scope.newProcedure).then(submitSuccess, submitError);
+    Procedures.submitProcedure($scope.procedure).then(submitSuccess, submitError);
   };
 
   var submitSuccess = function(response) {
     Messages.setMessage('Your procedure was created!', 'success');
+    $scope.procedure.title = "";
+    $scope.procedure.text = "";
+    $scope.showProcedure = false;
     $scope.getProcedures();
   };
 
@@ -30,38 +64,41 @@ angular.module('producer.procedures', ['alertMessage'])
     console.log('error: ', response);
   };
 
-  $scope.getProcedures = function(){
-    Procedures.getProcedures().then(function(resp){
-      $scope.procedures = resp.data;
-    });
+  $scope.updateProcedure = function() {
+    Procedures.updateProcedure($scope.procedure).then(updateSuccess, updateError);
+  };
+
+  var updateSuccess = function(response) {
+    Messages.setMessage('Your procedure was updated!', 'success');
+    $scope.getProcedures();
+  };
+
+  var updateError = function(response) {
+    Messages.setMessage('Sorry, there was an error updating your Procedure. Please try again.', 'error');
+    console.log('error: ', response);
+    $scope.getProcedures();
   };
 
   $scope.deleteProcedure = function(){
-    Procedures.deleteProcedure($scope.procedure).then(deleteSuccess, deleteError);
-  };
-
-  $scope.setMaster = function(section) {
-    $scope.selected = section;
-    $scope.newProcedure.title = $scope.selected.title;
-    $scope.newProcedure.text = $scope.selected.text;
-    $scope.showProcedure = true;
-    $scope.procedure = section;
-  };
-
-  $scope.createNewProcedure = function() {
-    $scope.newProcedure.title = "";
-    $scope.newProcedure.text = "";
-    $scope.showProcedure = false;
+    if ($scope.confirm) {
+      Procedures.deleteProcedure($scope.procedure).then(deleteSuccess, deleteError);
+      $scope.confirm = undefined;
+    } else {
+      $scope.confirm = 'content__delete--confirm';
+    }
   };
 
   var deleteSuccess = function(response) {
     Messages.setMessage('Procedure Deleted', 'success');
+    $scope.procedure = {};
+    $scope.showProcedure = false;
     $scope.getProcedures();
   };
 
   var deleteError = function(response) {
     Messages.setMessage('Sorry, there was an error submitting your form. Please submit again.', 'error');
     console.log('error: ', response);
+    $scope.getProcedures();
   };
 
   $scope.getProcedures();
